moved version switcher generation to scripts + slightly changed file structure for clarity
